+++ api/src/main/java/org/allaymc/api/world/WorldPool.java
@@ -13,17 +13,9 @@ import java.util.Map;
  */
 public interface WorldPool {
     /**
-     * Get the world settings of the server.
-     *
-     * @return the world settings of the server
-     */
-    WorldSettings getWorldConfig();
-
-    /**
      * Get the world by the name.
      *
      * @param name the name of the world
-     *
      * @return the world, or {@code null} if the world is not found
      */
     World getWorld(String name);
@@ -50,16 +42,14 @@ public interface WorldPool {
      *
      * @param name         the name of the world
      * @param worldSetting the settings of the world
-     *
      * @throws IllegalArgumentException if the world with the specific name already exists
      */
-    void loadWorld(String name, WorldSettings.WorldSetting worldSetting);
+    void loadWorld(String name, WorldSetting worldSetting);
 
     /**
      * Unload the world with the specific name.
      *
      * @param name the name of the world
-     *
      * @throws IllegalArgumentException if the world with the specific name is not exists, or the world is the default world
      */
     void unloadWorld(String name);